Move homopolymer helpers to utils
