+++ src/genecoder/gc_constrained_encoder.py
@@ -16,6 +16,7 @@ required.
 """
 
 from typing import Optional
+from .utils import check_homopolymer_length, get_max_homopolymer_length
 
 def calculate_gc_content(dna_sequence: str) -> float:
     """Calculates the GC content of a DNA sequence.
@@ -33,32 +34,6 @@ def calculate_gc_content(dna_sequence: str) -> float:
     gc_count = dna_sequence.upper().count('G') + dna_sequence.upper().count('C')
     return gc_count / len(dna_sequence)
 
-def check_homopolymer_length(dna_sequence: str, max_len: int) -> bool:
-    """Checks if any homopolymer in the DNA sequence exceeds a maximum length.
-
-    Args:
-        dna_sequence: The DNA sequence string.
-        max_len: The maximum allowed homopolymer length.
-
-    Returns:
-        True if any homopolymer is longer than max_len, otherwise False.
-        Returns False for an empty sequence.
-    """
-    if not dna_sequence:
-        return False
-
-    current_char = ''
-    current_len = 0
-    for char in dna_sequence.upper():
-        if char == current_char:
-            current_len += 1
-        else:
-            current_char = char
-            current_len = 1
-        
-        if current_len > max_len:
-            return True
-    return False
 
 def encode_gc_balanced(data: bytes, target_gc_min: float, target_gc_max: float, max_homopolymer: int) -> str:
     """Encodes binary data into a DNA sequence with GC content and homopolymer constraints.
@@ -100,36 +75,6 @@ def encode_gc_balanced(data: bytes, target_gc_min: float, target_gc_max: float, 
         # alternatives before falling back to this simple inversion.
         return "1" + alternative_sequence
 
-def get_max_homopolymer_length(dna_sequence: str) -> int:
-    """Calculates the length of the longest homopolymer in a DNA sequence.
-
-    Args:
-        dna_sequence: The DNA sequence string (e.g., "AAATTCGGGG").
-
-    Returns:
-        The length of the longest homopolymer. Returns 0 for an empty sequence.
-    """
-    if not dna_sequence:
-        return 0
-
-    max_len = 0
-    current_len = 0
-    if len(dna_sequence) > 0:
-        current_char = dna_sequence[0]
-        current_len = 1
-        max_len = 1
-
-    for i in range(1, len(dna_sequence)):
-        if dna_sequence[i] == current_char:
-            current_len += 1
-        else:
-            current_char = dna_sequence[i]
-            current_len = 1
-        
-        if current_len > max_len:
-            max_len = current_len
-            
-    return max_len if dna_sequence else 0
 
 
 def decode_gc_balanced(
