+++ tests/test_gc_constrained_encoder.py
@@ -7,10 +7,12 @@ from unittest.mock import patch, call  # call is needed for checking multiple ca
 from genecoder.encoders import encode_base4_direct  # noqa: E402
 from genecoder.gc_constrained_encoder import (
     calculate_gc_content,
-    check_homopolymer_length,
-    get_max_homopolymer_length,
     encode_gc_balanced,
     decode_gc_balanced,
+)
+from genecoder.utils import (
+    check_homopolymer_length,
+    get_max_homopolymer_length,
 )
 
 SRC_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src'))
